requirements.txt
delta-spark==3.0.0
pandas
pyarrow
matplotlib
seaborn
plotly
jupyterlab